pyproject.toml
[tool.poetry]
name            = "diffusion"
version         = "0.0.1"
description     = "Implementation of diffusion model"
readme          = "README.md"
license         = "MIT"
authors         = ["Justin Barry <justinbarrythejedi@gmail.com>"]

# tells Poetry which local package folder to install
packages = [
  { include = "diffusion" }
]

[tool.poetry.dependencies]
python          = ">=3.10,<4.0"
torch           = "==2.7.0"
torchvision     = "==0.22.0"
matplotlib      = "==3.10.3"
numpy           = "==2.2.6"
tqdm            = "==4.67.1"
omegaconf       = "==2.3.0"
wandb           = "==0.20.0"
torch-summary   = "==1.4.5"
torchinfo       = "==1.8.0"
antlr4-python3-runtime = "==4.9.3"
certifi         = "==2025.4.26"
charset-normalizer   = "==3.4.2"
click           = "==8.2.1"
contourpy       = "==1.3.2"
cycler          = "==0.12.1"
filelock        = "==3.18.0"
fonttools       = "==4.58.1"
fsspec          = "==2025.5.1"
gitdb           = "==4.0.12"
gitpython       = "==3.1.44"
idna            = "==3.10"
jinja2          = "==3.1.6"
kiwisolver      = "==1.4.8"
markupsafe      = "==3.0.2"
mpmath          = "==1.3.0"
networkx        = "==3.4.2"
packaging       = "==25.0"
pillow          = "==11.2.1"
platformdirs    = "==4.3.8"
protobuf        = "==6.31.1"
psutil          = "==7.0.0"
pydantic        = "==2.11.5"
pydantic-core   = "==2.33.2"
pyparsing       = "==3.2.3"
python-dateutil = "==2.9.0.post0"
pyyaml          = "==6.0.2"
requests        = "==2.32.3"
sentry-sdk      = "==2.29.1"
setproctitle    = "==1.3.6"
six             = "==1.17.0"
smmap           = "==5.0.2"
sympy           = "==1.14.0"
typing-inspection = "==0.4.1"
typing-extensions  = "==4.14.0"
urllib3         = "==2.4.0"
gdown           = ">=5.2.0,<6.0.0"
lpips           = ">=0.1.4,<0.2.0"
transformers = "^4.53.0"

[tool.poetry.dev-dependencies]
setuptools      = ">=80.9.0,<81.0.0"

[build-system]
requires        = ["poetry-core>=2.0.0,<3.0.0"]
build-backend   = "poetry.core.masonry.api"